Modularize Current Code
- lib.rs was getting pretty long, so I made some modules

- Even though they're small right now they will surely grow enough
  to warrant their own modules
